Изменил названия файлов, и сложил их в отдельный модуль. Вне модуля вынес main.py
- Нижний регистр MESSAGES.py;
- Объединил bot.py и commands.py. Отделять логику handler'ов от логики самих команд показалось избыточным.
- core.py описывает создание главных компонентов чат-бота;
- __init__.py задает общую область видимости внутри модуля (папки bot_app);

- Поменял PyMySQL на mysql-connector-python -- официальную реализацию API для работы с MySQL. Так вышло стабильнее.
- Пример с выводом данных преподавателей по их ID работает корректно.
