--- MESSAGES.py
@@ -1,4 +1,0 @@
-TODAY = "Занятия сегодня"
-TOMOROW = "Занятия завтра"
-THIS_WEEK = "Занятия на этой неделе"
-NEXT_WEEK = "Занятия на следующей неделе"
