+++ bot_app/__init__.py
@@ -1,0 +1,9 @@
+"""
+Модуль чат-бота
+"""
+import logging
+
+from .core import dp
+from . import commands
+
+logging.basicConfig(level=logging.INFO)
