+++ bot_app/messages.py
@@ -1,0 +1,9 @@
+"""
+Текст сообщений, используемый в ответах чат-бота на запросы пользователя
+"""
+
+TODAY = "Занятия сегодня"
+TOMOROW = "Занятия завтра"
+THIS_WEEK = "Занятия на этой неделе"
+NEXT_WEEK = "Занятия на следующей неделе"
+GREETINGS = "Привет!\nВведи число и получишь фамилию одного из преподавателей"
